// Neptune.Web/src/app/shared/generated/model/treatment-bmp-benchmark-and-threshold-with-observation-type-dto.ts
export class TreatmentBMPBenchmarkAndThresholdWithObservationTypeDto { 
    TreatmentBMPBenchmarkAndThresholdID?: number | null;
    TreatmentBMPID?: number;
    TreatmentBMPTypeID?: number;
    TreatmentBMPTypeAssessmentObservationTypeID?: number;
    TreatmentBMPAssessmentObservationTypeID?: number;
    ObservationTypeName?: string | null;
    BenchmarkUnitLabel?: string | null;
    ThresholdUnitLabel?: string | null;
    BenchmarkValue?: number | null;
    ThresholdValue?: number | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}